requirements.txt
requests>=2.18.4,<3.0
google-auth>=1.0,<2.0
